Implement equilibrium task callback
This is the implementation of the first task handling a request for an
equilibrium job. Although the code seems more complicated than celery,
it enables the sort of very complex workflows that we require in this
appliucation. To summarize the operation (this will go in the docs as
well):

1. Receive the incoming body, which is bytes
2. unpack the components of that that will be required to run an
equilibrium task
3. Call the equilibrium task
4. Publish different subsets of the results to different topics

In particular, this enables a few desired things:

1. We can route different parts of the results to different tasks,
preventing the need to match returns and signatures
2. We can actually have cycles in the compute graph. This function will
call itself by sending a message to its own topic
3. It enables listeners to capture whatever information they want on
whatever topic they want without disturbing the calculation.

It should be noted that this expects all the information necessary to
run both an equilibrium and a nonequilibrium task. While we may be able
to generate the necessary nonequilibrium stuff in this task, it seems a)
messy and b) would still require additional flags.