Begin adding writer worker
This module implements a separate writer worker. There are a couple
reasons why we might want the writer to be separate:

1) It can run on non-GPU resources, and avoid competing with GPU
resources

2) We can use the rabbitmq routing system to direct writes for different
projects to different locations

3) We can run it locally with everything else running on the
cluster/cloud, and collect just the outputs (without inadvertently
adding the local machine as a compute worker).

As a TODO, this module still needs the nonequilibrium callback to be
implemented, and also needs a basic command line input to be implemented
(this is so that we can select a particular project, set of projects, or
(default) all projects for writing).